// src/types/widget.tsx
export interface widgetPropsI {
  title: string;
  content: string;
  type: string;
}

export interface widgetColorsI {
  body_color?: string;
  bubble_background_color?: string;
  cta_color?: string;
  cta_font_color?: string;
  font_color?: string;
  heading_background_color?: string;
  heading_font_color?: string;
  transcript_color?: string;
}